+++ packages/components/src/flowNavigation/FlowNavigation.js
@@ -23,8 +23,8 @@ const BackArrow = () => (
   </svg>
 );
 
-const FlowNavigation = ({ steps, activeStep, avatarUrl, profileType, onClose, onGoBack }) => (
-  <div className="tw-flow-navigation">
+const FlowNavigation = ({ steps, activeStep, avatarUrl, done, profileType, onClose, onGoBack }) => (
+  <div className={`tw-flow-navigation ${done ? 'tw-flow-navigation--done' : ''}`}>
     <div className="container">
       <div className="row p-t-3 tw-flow-navigation__wrapper">
         <div className="col-lg-2 col-xs-6 m-lg-t-1">
@@ -40,15 +40,20 @@ const FlowNavigation = ({ steps, activeStep, avatarUrl, profileType, onClose, on
           <div className={`flag flag-info logo-3 visible-xs ${onGoBack ? 'flag--hidden' : ''}`} />
         </div>
         <div className="col-lg-2 col-xs-6 col-lg-push-8 text-xs-right m-lg-t-1">
-          <Avatar url={avatarUrl} profileType={profileType} />
+          {done ? '' : <Avatar url={avatarUrl} profileType={profileType} />}
           <button
-            className="btn-unstyled tw-flow-navigation__close-button icon icon-close icon-lg m-l-3"
+            className={`btn-unstyled tw-flow-navigation__close-button
+              ${done ? '' : 'close-button-with-avatar'} icon icon-close icon-lg m-l-3`}
             onClick={onClose}
           />
         </div>
-        <div className="col-xs-12 col-lg-6 col-lg-pull-2 col-lg-offset-1 tw-flow-navigation__stepper">
-          <Stepper activeStep={activeStep} steps={steps} />
-        </div>
+        {done ? (
+          ''
+        ) : (
+          <div className="col-xs-12 col-lg-6 col-lg-pull-2 col-lg-offset-1 tw-flow-navigation__stepper">
+            <Stepper activeStep={activeStep} steps={steps} />
+          </div>
+        )}
       </div>
     </div>
   </div>
@@ -72,6 +77,7 @@ FlowNavigation.propTypes = {
   ).isRequired,
   activeStep: Types.number,
   avatarUrl: Types.string,
+  done: Types.bool,
   profileType: Types.oneOf(Object.keys(ProfileType)),
   onGoBack: Types.func,
   onClose: Types.func.isRequired,